+++ packages/blunderbuss/src/utils.ts
@@ -15,7 +15,7 @@
 // eslint-disable-next-line node/no-extraneous-import
 import {Context} from 'probot';
 import * as util from 'util';
-import {logger} from 'gcf-utils';
+import {logger as defaultLogger, GCFLogger} from 'gcf-utils';
 import {
   PullRequestEvent,
   IssuesEvent,
@@ -62,7 +62,8 @@ export function isIssue(
 
 export async function assign(
   context: Context<'issues'> | Context<'pull_request'>,
-  config: Configuration
+  config: Configuration,
+  logger: GCFLogger = defaultLogger
 ) {
   let issue: Issue | undefined;
   let pullRequest: PullRequest | undefined;