+++ packages/sync-repo-settings/src/bot.ts
@@ -14,7 +14,7 @@
 
 // eslint-disable-next-line node/no-extraneous-import
 import {Probot, Context} from 'probot';
-import {logger} from 'gcf-utils';
+import {getContextLogger} from 'gcf-utils';
 import {ConfigChecker, getConfig} from '@google-automations/bot-config-utils';
 
 import {SyncRepoSettings} from './sync-repo-settings';
@@ -55,6 +55,7 @@ export function handler(app: Probot) {
    * If so, run the settings sync for that repo.
    */
   app.on('push', async context => {
+    const logger = getContextLogger(context);
     const branch = context.payload.ref;
     const defaultBranch = context.payload.repository.default_branch;
     if (branch !== `refs/heads/${defaultBranch}`) {
@@ -102,6 +103,7 @@ export function handler(app: Probot) {
 
   // eslint-disable-next-line @typescript-eslint/no-explicit-any
   app.on(['schedule.repository' as any], async context => {
+    const logger = getContextLogger(context);
     logger.info(`running for org ${context.payload.cron_org}`);
     const {owner, repo} = context.repo();
     if (context.payload.cron_org !== owner) {
@@ -125,6 +127,7 @@ export function handler(app: Probot) {
   });
 
   app.on('repository.transferred', async context => {
+    const logger = getContextLogger(context);
     const {owner, repo} = context.repo();
     const defaultBranch = context.payload.repository.default_branch;
     const config = await getConfig<RepoConfig>(